+++ g2g-plugins/src/mfdecode.rs
@@ -376,7 +376,6 @@ impl AsyncElement for MfDecode {
                 }
                 let frame = Frame {
                     domain: MemoryDomain::System(SystemSlice::from_boxed(d.bytes)),
-                    caps: new_caps,
                     timing: FrameTiming {
                         pts_ns: d.pts_ns,
                         dts_ns: d.pts_ns,